Use npm@3.3.9 to manage npm dependencies of packages.
We were already using npm@3.1.2 to install the contents of
dev_bundle/{lib,server-lib}/node_modules, with care to restore the
original npm@1.4.28 before bundling. This commit uses npm@1.4.28 to
install the latest version of npm (3.3.9) over the original version.

Getting `meteor test-packages` to run and pass required some very minor
changes in tools/isobuild/meteor-npm.js, specifically not deleting the
package.json and node-shrinkwrap.json files that we create in .npm/package
directories.

It makes sense that npm would be so backwards compatible, since it is
constrained to deliver the same files for a particular version of a
package as it ever did.