Remove clientState from API in favor of using existing mechanisms
Instead of having `clientState` explicitly passed around in the identity system,
code which calls `Identity.create` or `Identity.authenticate` should first
ensure that any client state is stored in a way which will be preserved by the
identity service(s) it uses. When using services like password, the state
doesn't require any special handling. When using redirecting oauth services, the
state should be stored in a named `ReactiveDict`.  When using services which
might complete on a different client (e.g. passwordless email/sms only
services), the state should be stored in the URL (e.g. via whatever routing
package is in use). A general purpose accounts UI package should always store
the state in the URL to ensure it is preserved when using all services.

This change is driven by the realization that the "preserve client state" story
is orthogonal to accounts/identities and already has existing solutions like
named `ReactiveDict` and the URL management provided by router packages.